Restructure project to make things more clear
