+++ cmd/main.go
@@ -21,8 +21,8 @@ import (
 	"github.com/go-logr/zapr"
 	kodev1alpha1 "github.com/jacero-io/kode-operator/api/v1alpha1"
 	"github.com/jacero-io/kode-operator/internal/cleanup"
-	entrypointcontroller "github.com/jacero-io/kode-operator/internal/controller"
-	kodecontroller "github.com/jacero-io/kode-operator/internal/controller"
+	entrypointcontroller "github.com/jacero-io/kode-operator/internal/controllers/entrypoint"
+	kodecontroller "github.com/jacero-io/kode-operator/internal/controllers/kode"
 	"github.com/jacero-io/kode-operator/internal/resource"
 	"github.com/jacero-io/kode-operator/internal/status"
 	"github.com/jacero-io/kode-operator/internal/template"
